backend/requirements.txt
alembic==1.16.1
amqp==5.3.1
aniso8601==10.0.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
billiard==4.2.1
black==23.7.0
blinker==1.9.0
Brotli==1.1.0
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.8.2
cssselect2==0.8.0
Deprecated==1.2.18
distlib==0.3.9
distro==1.9.0
filelock==3.18.0
flake8==6.0.0
Flask==2.3.3
flask-cors==6.0.0
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.12
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
fonttools==4.58.0
fsspec==2025.5.1
greenlet==3.2.2
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.0
identify==2.6.12
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
isort==5.12.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.5.3
limits==5.2.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==2.2.6
openai==1.82.0
ordered-set==4.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
pre-commit==3.3.3
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycodestyle==2.10.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydyf==0.11.0
pyflakes==3.0.1
Pygments==2.19.1
PyJWT==2.10.1
pyphen==0.17.2
pytest==7.4.0
pytest-cov==4.1.0
pytest-flask==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.1
requests==2.32.3
rich==13.9.4
rpds-py==0.25.1
safetensors==0.5.3
sentry-sdk==2.29.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
structlog==25.3.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.52.3
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
virtualenv==20.31.2
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
zopfli==0.2.3.post1
